Backend/Lambda_New/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.6.0+cpu
torchvision==0.21.0+cpu
pillow==11.3.0
